+++ src/fs.rs
@@ -192,17 +192,16 @@ pub unsafe fn nstd_fs_remove_dirs(name: &NSTDStr) -> NSTDIOError {
 ///
 /// `NSTDIOBufferResult contents` - The file's contents, or the I/O operation error code on failure.
 ///
-/// # Panics
-///
-/// This operation will panic if allocating fails.
-///
 /// # Safety
 ///
 /// This operation can cause undefined behavior if `path`'s data is invalid.
 #[nstdapi]
 pub unsafe fn nstd_fs_read(path: &NSTDStr) -> NSTDIOBufferResult {
     match std::fs::read(path.as_str()) {
-        Ok(contents) => NSTDResult::Ok(NSTDVec::from_slice(&contents)),
+        Ok(contents) => match NSTDVec::from_slice(&contents) {
+            NSTDOptional::Some(contents) => NSTDResult::Ok(contents),
+            _ => NSTDResult::Err(NSTDIOError::NSTD_IO_ERROR_OUT_OF_MEMORY),
+        },
         Err(err) => NSTDResult::Err(NSTDIOError::from_err(err.kind())),
     }
 }
@@ -217,17 +216,16 @@ pub unsafe fn nstd_fs_read(path: &NSTDStr) -> NSTDIOBufferResult {
 ///
 /// `NSTDIOStringResult contents` - The file's contents, or the I/O operation error code on failure.
 ///
-/// # Panics
-///
-/// This operation will panic if allocating fails.
-///
 /// # Safety
 ///
 /// This operation can cause undefined behavior if `path`'s data is invalid.
 #[nstdapi]
 pub unsafe fn nstd_fs_read_to_string(path: &NSTDStr) -> NSTDIOStringResult {
     match std::fs::read_to_string(path.as_str()) {
-        Ok(contents) => NSTDResult::Ok(NSTDString::from_str(&contents)),
+        Ok(contents) => match NSTDString::from_str(&contents) {
+            NSTDOptional::Some(contents) => NSTDResult::Ok(contents),
+            _ => NSTDResult::Err(NSTDIOError::NSTD_IO_ERROR_OUT_OF_MEMORY),
+        },
         Err(err) => NSTDResult::Err(NSTDIOError::from_err(err.kind())),
     }
 }
@@ -319,10 +317,6 @@ pub unsafe fn nstd_fs_copy(from: &NSTDStr, to: &NSTDStr) -> NSTDIOError {
 /// `NSTDIOStringResult contents` - The absolute version of `path`, or the I/O operation error code
 /// on failure.
 ///
-/// # Panics
-///
-/// This operation will panic if allocating fails.
-///
 /// # Safety
 ///
 /// This operation can cause undefined behavior if `path`'s data is invalid.
@@ -330,7 +324,10 @@ pub unsafe fn nstd_fs_copy(from: &NSTDStr, to: &NSTDStr) -> NSTDIOError {
 pub unsafe fn nstd_fs_absolute(path: &NSTDStr) -> NSTDIOStringResult {
     match std::fs::canonicalize(path.as_str()) {
         Ok(path) => match path.into_os_string().into_string() {
-            Ok(path) => NSTDResult::Ok(NSTDString::from_str(&path)),
+            Ok(path) => match NSTDString::from_str(&path) {
+                NSTDOptional::Some(path) => NSTDResult::Ok(path),
+                _ => NSTDResult::Err(NSTDIOError::NSTD_IO_ERROR_OUT_OF_MEMORY),
+            },
             _ => NSTDResult::Err(NSTDIOError::NSTD_IO_ERROR_INVALID_DATA),
         },
         Err(err) => NSTDResult::Err(NSTDIOError::from_err(err.kind())),